fixed enemy movement and spawning
